A helper function to add treatment data to a chronogram

A helper function to add treatment data to a chronogram

Usage

cg_add_treatment(cg, treatment)

Arguments

cg

a chronogram object (class cg_tbl)

treatment

a tibble containing treatment dates (with columns: the specified calendar_date, and ID columns, defined with chronogram_skeleton())

This code wraps thecg_add_experiment function.

Value

A chronogram

See also

Examples

if (FALSE) { # \dontrun{

## Create a tibble of exemplar treatment data ##
# Here, we have treated ID=1 with B cell depletion
# therapy, rituximab (RTX).
#
small_study_treatment <-
  tibble::tribble(
    ~elig_study_id, ~calendar_date, ~treatment,
    1, "01/02/2021", "RTX"
  )
small_study_treatment <- small_study_treatment %>%
  mutate(across(contains("date"), ~ lubridate::dmy(.x)))

## Add to chronogram ##
small_study_chronogram <- cg_add_treatment(
  small_study_chronogram,
  small_study_treatment
)
} # }