Cumulative counting seroconversion episodes per individual

cg_annotate_exposures_count returns a cumulative count of exposures. These exposures can be vaccines, infection episodes, or seroconversion episodes (a seroconversion episode is biologically an infection - the difference between the two is concomittant symptoms, or diagnostic testing).

Usage

cg_annotate_exposures_count(
  x,
  episode_number = episode_number,
  N_seroconversion_episode_number = N_seroconversion_episode_number,
  dose_number = dose_number,
  exposure_number = exposure_number
)

Arguments

x

a chronogram

episode_number

the column containing infection episode numbers. cg_annotate_episodes_find() Default is episode_number.

N_seroconversion_episode_number

the column name for resulting episode flag. Generated by cg_annotate_episodes_find_seroconversion(). Default is "N_seroconversion_episode_number".

dose_number

the column name for cumulative doses. Generated by cg_annotate_vaccines_count(). Default is "dose_number".

exposure_number

the column name to return the cumulative counter. Default is "exposure_number".

Value

x a chronogram, with episode numbers annotated

See also

Examples

if (FALSE) { # \dontrun{

data("built_smallstudy")
cg <- built_smallstudy$chronogram

## Simulate some infection data ##
infections_to_add <- tibble::tribble(
  ~calendar_date, ~elig_study_id, ~LFT, ~PCR, ~symptoms,
  "01102020", "1", "pos", NA, NA,
  "11102020", "1", "pos", NA, "severe"
)
## Make calendar_date a date ##
infections_to_add$calendar_date <- lubridate::dmy(
  infections_to_add$calendar_date
)
## add to chronogram
cg <- cg_add_experiment(cg, infections_to_add)

## annotate vaccines ##
cg <- cg_annotate_vaccines_count(cg,
  dose = dose,
  dose_counter = dose_number,
  vaccine_date_stem = date_dose,
  intermediate_days = 7
)

## now infection finding ##
cg <- cg_annotate_episodes_find(cg,
  infection_cols = c("LFT", "PCR", "symptoms"),
  infection_present = c("pos", "Post", "^severe")
)

cg <- cg_annotate_episodes_find_seroconversion(cg,
  serum_N_titre =
    "serum_Ab_N"
)

cg.final <- cg_annotate_exposures_count(cg)
} # }