Calculate and assign flags based on infection history

Calculate and assign flags based on infection history

Usage

cg_annotate_antigenic_history(
  x,
  episode_number = episode_number,
  dose_number = dose_number,
  episode_variant_summarised = episode_variant_summarised,
  ag_col = antigenic_history
)

Arguments

x

a chronogram

episode_number

a character vector to identify the episode_number column. Default is "episode_number".

dose_number

a character vector to identify the column containing the number of doses received on that day. Default is "dose_number".

episode_variant_summarised

a character vector to identify the column summarising the variant call. Default is "episode_variant_summarised".

ag_col

column with infection flag

Value

A chronogram

Examples

if (FALSE) { # \dontrun{
cg <- cg_annotate_label_episodes(cg)
} # }