Calculate and assign flags based on infection history

This counts the total number of infections for each individual at the end of the study (or at that data freeze), and makes it available as a fresh column. this function should be run after cg_annotate_episodes_find(). For the cumulative number of infections use cg_annotate_episodes_find() only.

Usage

cg_annotate_episodes_count(
  x,
  episode_number = episode_number,
  count_col = count_col
)

Arguments

x

a chronogram

episode_number

a character vector to use to label the ID column. Default is "episode_number".

count_col

column with infection count

Value

A subsetted chronogram

See also

Examples

if (FALSE) { # \dontrun{
cg <- cg_annotate_episodes_count(cg)
} # }