Find infections per individual

Find infections per individual

Usage

cg_annotate_episodes_fill(
  x,
  col_to_fill,
  col_to_return,
  .direction = "down",
  episode_numbers_col = episode_number
)

Arguments

x

a chronogram

col_to_fill

the name of the column to fill within each episode

col_to_return

the name of the returned, filled column

.direction

See tidyr::fill(), provide quoted.

episode_numbers_col

The column name to use for episode numbers. Default is episode_number, unquoted.

Value

x a chronogram, with episode data filled within each episode.

Examples

if (FALSE) { # \dontrun{

## Simulate some infection data ##
infections_to_add <- tibble::tribble(
  ~calendar_date, ~elig_study_id, ~LFT, ~PCR, ~symptoms,
  "01102020", "1", "pos", NA, NA,
  "11102020", "1", "pos", NA, "severe"
)
## Make calendar_date a date ##
infections_to_add$calendar_date <- lubridate::dmy(
  infections_to_add$calendar_date
)
## add to chronogram
small.study <- cg_add_experiment(small.study, infections_to_add,
  ids_column_name = "elig_study_id",
  calendar_date = "calendar_date"
)

## now infection finding ##
small.study.inf <- cg_annotate_episodes_find(small.study,
  infection_cols = c("LFT", "PCR", "symptoms"),
  infection_present = c("pos", "Post", "^severe")
)
summary(small.study.inf$episode_number)

## exact text matching ##
test2 <- cg_annotate_episodes_find(small.study,
  infection_cols = c("LFT", "PCR", "symptoms"),
  infection_present = c("Pos", "Post", "^mild")
)
summary(test2$episode_number)

## empty strings will error (as they otherwise match everything) ##
test3a <- cg_annotate_episodes_find(small.study,
  infection_cols = c("LFT", "PCR", "symptoms"),
  infection_present = c("pos", "Post", "")
)
## a 'random' string will not error ##
test3b <- cg_annotate_episodes_find(small.study,
  infection_cols = c("LFT", "PCR", "symptoms"),
  infection_present = c("pos", "Post", "a")
)

summary(test2$episode_number)
} # }